requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
paramiko==3.3.1
redis==5.0.1
telnetlib3==2.0.2
python-dotenv==1.0.0
aioredis==2.0.1
asyncio-mqtt==0.16.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
asyncpg==0.29.0
httpx==0.25.2
loguru==0.7.2
asyncssh==2.14.1